Piccole modifiche di formattazione
Piccole modifiche di formattazione

+++ src/main/java/it/eagleprojects/progettoverificamybatis4/dao/CorsoMapper.java
@@ -14,21 +14,21 @@ import it.eagleprojects.progettoverificamybatis4.model.Corso;
 
 @Repository
 public class CorsoMapper {
-	
+
 	/**
 	 * Metodo che ritorna tutte le righe della tabella corsi
 	 * @return una <code> List </code> di Corso
 	 */
 	public List<Corso> getAllCorsi(){
-		
+
 		SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession();
 		List<Corso> listaCorsi = session.selectList("getAllCorsi"); //Qui il nome è quello definito nel file mapper xml
 		session.commit();
 		session.close();
 		return listaCorsi;
 	}
-	
-	
+
+
 	/**
 	 * Metodo che ritorna una riga corrispondente al corsoId specificato come parametro
 	 * @return una istanza di Corso
@@ -40,26 +40,26 @@ public class CorsoMapper {
 		session.close();
 		return corso;
 	}
-	
-	
-	
+
+
+
 	/**
 	 * Metodo che ritorna tutti i corsi associati allo studenteId specificato come parametro
 	 * @return una <code> List </code> di Corso
 	 */
 	public List<Corso> getAllCorsiByStudenteId(@Param("studenteId") long studenteId){
-		
+
 		SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession();
 		List<Corso> corsiList = session.selectList("getAllCorsiByStudenteId", studenteId); //Qui il nome è quello definito nel file mapper xml
 		session.commit();
 		session.close();
 		return corsiList;
 	}
-	
-	
-	
+
+
+
 	/**
-	 *  Metodo che salva una istanza di <code> Corso </code> nel database sottostante 
+	 *  Metodo che salva una istanza di <code> Corso </code> nel database sottostante
 	 * @param e l'istanza da persistere
 	 */
 	public Corso saveCorso(Corso corso) {
@@ -69,7 +69,7 @@ public class CorsoMapper {
 		session.close();
 		return corso;
 	}
-	
+
 	/**
 	 *  Metodo che aggiunge una istanza di <code> Corso </code> allo Studente specificato
 	 * @param e l'istanza da aggiungere
@@ -81,43 +81,43 @@ public class CorsoMapper {
 		session.commit();
 		session.close();
 	}
-	
-	
-	
+
+
+
 	/**
-	 * Metodo che aggiorna una istanza di <code> Corso </code> nel database sottostante 
+	 * Metodo che aggiorna una istanza di <code> Corso </code> nel database sottostante
 	 * @return una istanza di Corso (quella appena aggiornata)
 	 */
 	public Corso updateCorsoById(@Param("corso") Corso corso) {
 		SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession();
-		int righeModificate = (int) session.update("updateCorsoById", corso); //Qui il nome è quello definito nel file mapper xml
+		int righeModificate = session.update("updateCorsoById", corso); //Qui il nome è quello definito nel file mapper xml
 		session.commit();
 		session.close();
 		return corso;
 	}
-	
+
 	/**
-	 * Metodo che elimina una istanza di <code> Corso </code> nel database sottostante 
+	 * Metodo che elimina una istanza di <code> Corso </code> nel database sottostante
 	 * @return void
 	 */
 	public void deleteCorsoById(@Param("corsoId") Long corsoId) {
 		SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession();
-		int righeModificate = (int) session.delete("deleteCorsoById", corsoId); //Qui il nome è quello definito nel file mapper xml
+		int righeModificate = session.delete("deleteCorsoById", corsoId); //Qui il nome è quello definito nel file mapper xml
 		session.commit();
 		session.close();
 	}
-	
+
 	/**
-	 * Metodo che elimina tutte le istanze <code> Corso </code> nel database sottostante 
+	 * Metodo che elimina tutte le istanze <code> Corso </code> nel database sottostante
 	 * @return void
 	 */
 	public void deleteAllCorsi() {
 		SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession();
-		int righeModificate = (int) session.delete("deleteAllCorsi"); //Qui il nome è quello definito nel file mapper xml
+		int righeModificate = session.delete("deleteAllCorsi"); //Qui il nome è quello definito nel file mapper xml
 		session.commit();
 		session.close();
 	}
-	
+
 	/**
 	 *  Metodo che elimina una istanza di <code> Corso </code> dallo Studente specificato
 	 * @param e l'istanza da aggiungere
